feat: allow setting backend url at build time
Vite build requires knowing about the backend url. Previously, this was
hardcoded in the Dockerfile as VITE_REACT_APP_BACKEND_URL. This change
allows setting this variable through BACKEND_URL build argument.

It is used with --build-arg option in docker build command.
The name of the variable has been changed for a shorter one, easier to
type and remember.