Changes to Dockerfile and documentation (#115)
* feat: allow setting backend url at build time

Vite build requires knowing about the backend url. Previously, this was
hardcoded in the Dockerfile as VITE_REACT_APP_BACKEND_URL. This change
allows setting this variable through BACKEND_URL build argument.

It is used with --build-arg option in docker build command.
The name of the variable has been changed for a shorter one, easier to
type and remember.

* capitalize the filenames of readme.md and license.txt

Remove the .txt extension of LICENSE file, too. This change is to align
with the accepted standards for these files.

* improve README

* give instructions to build docker image with BACKEND_URL build arg
* move the sentence about the origin of the project at the bottom

* removed the need for trailing slash in VITE_REACT_APP_BACKEND_URL, see #115

* properly handle trailing slash

* ommit module from Dockerignore

---------

Co-authored-by: Manuel Lera-Ramirez <manulera14@gmail.com>